slå sammen config og oppstart
